move AddServerIcon to AddServerComponent folder, create seperate file for customSVG and style
