+++ src/components/icons/icons.module.css
@@ -85,86 +85,3 @@ p {
     content: none;
 }
 
-/* add server */
-
-.addServerIcon {
-    height: 7vh;
-    width: 7vh;
-    background-color: #9d37a873;
-    border-radius: 50%;
-    margin: 5px 5px 0 5px;
-    cursor: pointer;
-    transition: all 0.2s ease-in-out;
-    border-color: transparent;
-}
-
-.addServerIcon:hover {
-    background-color: #1d68d0b2;
-    border-radius: 30%;
-}
-
-.addServerIconImage {
-    margin-top: 15px;
-}
-
-.addServerModalHeader {
-    padding: 24px 24px 0px 24px;
-    border: none;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-}
-
-.addServerModalHeader h1 {
-    text-align: center;
-    font-size: 23px;
-    font-weight: 700;
-}
-
-.addServerModalHeader div {
-    text-align: center;
-    font-size: 16px;
-    width: 90%;
-    color: #9a9494;
-}
-
-.addServerModalBody {
-    border: none;
-    margin-top: 16px;
-    margin-bottom: 16px;
-    padding-left: 16px;
-    padding-right: 16px;
-    height: 200px;
-}
-
-.addServerModalBodyContent {
-    width: 100%;
-    display: flex;
-    justify-content: center;
-}
-
-.addServerModalBodyFormHeader {
-    margin: 24px 0px 8px 0px;
-    width: 100%;
-    font-size: 13px;
-    font-weight: 700;
-    color: rgb(102, 100, 100);
-}
-
-.addServerModalBodyFormContent {
-    border: none;
-    background-color: #f3efef;
-    border-radius: 6px;
-    width: 100%;
-    height: 45px;
-}
-
-.addServerModalFooter {
-    border: none;
-    background-color: #f3efef;
-}
-
-.addServerModalFooterButton {
-    background-color: #516eda;
-    width: 22%;
-}
