+++ src/components/nav/Nav.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import GroupIcon from "../icons/GroupIcon";
-import AddServerIcon from '../icons/AddServerIcon';
+import AddServerIcon from '../AddServerComponents/AddServerIcon';
 import styles from './nav.module.css'
 import { useNavigate } from "react-router-dom";
 import { getGroups } from '../../firebase';
